requirements.txt
pydantic-ai[all]>=0.0.14
gradio>=5.9.1
httpx>=0.28.1
python-dotenv>=1.0.1
langfuse>=2.54.4
pydantic>=2.10.3
python-multipart>=0.0.12
uvicorn[standard]>=0.32.1
pytest>=8.3.4
pytest-asyncio>=0.24.0
black>=24.10.0
click>=8.0.0
pillow>=10.0.0
pyautogui>=0.9.54